+++ board_files/include/classes/BanHammer.php
@@ -101,7 +101,7 @@ class BanHammer
         if (!$this->authorize->get_user_perm($_SESSION['username'], 'perm_ban_add', INPUT_BOARD_ID) &&
              !$authorize->get_user_perm($_SESSION['username'], 'perm_all_ban_modify'))
         {
-            nel_derp(321, nel_stext('You are not allowed to add new bans.'));
+            nel_derp(321, _gettext('You are not allowed to add new bans.'));
         }
 
         $prepared = $this->dbh->prepare('INSERT INTO "' . BAN_TABLE . '" ("board_id", "all_boards", "type", "ip_address_start", "reason", "length", "start_time")
@@ -130,7 +130,7 @@ class BanHammer
         if (!$this->authorize->get_user_perm($_SESSION['username'], 'perm_ban_modify', INPUT_BOARD_ID) &&
              !$authorize->get_user_perm($_SESSION['username'], 'perm_all_ban_modify'))
         {
-            nel_derp(322, nel_stext('You are not allowed to modify bans.'));
+            nel_derp(322, _gettext('You are not allowed to modify bans.'));
         }
 
         $prepared = $this->dbh->prepare('UPDATE "' . BAN_TABLE .
@@ -156,7 +156,7 @@ class BanHammer
             if (!$this->authorize->get_user_perm($_SESSION['username'], 'perm_ban_delete', INPUT_BOARD_ID) &&
                  !$this->authorize->get_user_perm($_SESSION['username'], 'perm_all_ban_modify') && !$snacks)
             {
-                nel_derp(323, nel_stext('You are not allowed to delete bans.'));
+                nel_derp(323, _gettext('You are not allowed to delete bans.'));
             }
         }