+++ board_files/include/classes/FileHandler.php
@@ -133,7 +133,7 @@ class FileHandler
 
         if ($filtered === '')
         {
-            nel_derp(111, nel_stext('Filename was empty or was purged by filter.'));
+            nel_derp(111, _gettext('Filename was empty or was purged by filter.'));
         }
 
         return $filtered;