+++ board_files/include/output/rules.php
@@ -47,7 +47,7 @@ function nel_render_rules_list($board_id)
     $thumbnail_limit = $rules_item_element->cloneNode(true);
     $thumbnail_limit->firstChild->setContent('Images greater than ');
     $thumbnail_limit->firstChild->nextSibling->setContent($board_settings['max_width'] . ' x ' .
-         $board_settings['max_height'] . nel_stext(' pixels will be thumbnailed.'));
+         $board_settings['max_height'] . _gettext(' pixels will be thumbnailed.'));
     $post_limits->appendChild($thumbnail_limit);
 
     $rules_div->appendChild($filetype_rules);